chore(react): Move examples to react-onsenui-examples app
