--- bindings/react/examples/index.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8" />
-  <link href='https://fonts.googleapis.com/css?family=Roboto:400,300,300italic,400italic,500,500italic,700,700italic' rel='stylesheet' type='text/css'>
-  <link rel="stylesheet" href="../../../build/css/onsenui.css">
-  <link rel="stylesheet" href="../../../build/css/onsen-css-components.css">
-
-  <title>Onsen UI React Demo</title>
-
-  <style>
-    ons-col {
-      border: 1px solid #ccc;
-      background: #fff;
-      overflow: hidden;
-      padding: 4px;
-      color: #999;
-    }
-
-    .page__content {
-      background-color: #f6f6f6;
-    }
-
-    .dialog {
-      height: 300px !important;
-    }
-
-    h3 {
-      color: #666;
-      font-size: 17px;
-    }
-  </style>
-</head>
-
-<body>
-  <div id="app" >
-  </div>
-  <script src="build.js"></script>
-</body>
-</html>
