+++ bindings/react/react-onsenui-examples/src/App.js
@@ -1,25 +1,38 @@
-import logo from './logo.svg';
-import './App.css';
+import React from 'react';
+import ons from 'onsenui';
+import { Navigator } from 'react-onsenui';
+import Examples from './Examples';
 
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
+import 'onsenui/css/onsenui.css';
+import 'onsenui/css/onsen-css-components.css';
+
+class App extends React.Component {
+  renderPage(route, navigator) {
+    const props = route.props || {};
+    props.navigator = navigator;
+
+    return React.createElement(route.component, route.props);
+  }
+
+  render() {
+    ons.mockStatusBar();
+    return (
+      <Navigator
+        swipeable={true}
+        renderPage={this.renderPage}
+        onPrePush={e => console.log('prepush', e)}
+        onPostPush={e => console.log('postpush', e)}
+        onPrePop={e => console.log('prepop', e)}
+        onPostPop={e => console.log('postpop', e)}
+        initialRoute={{
+          component: Examples,
+          props: {
+            key: 'examples'
+          }
+        }}
+      />
+    );
+  }
 }
 
 export default App;
