+++ bindings/react/react-onsenui-examples/src/Examples.js
@@ -1,11 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-
-import ons from 'onsenui';
-
 import {
   Page,
-  Navigator,
   Toolbar,
   List,
   ListItem
@@ -52,33 +47,4 @@ class Examples extends React.Component {
   }
 }
 
-class App extends React.Component {
-  renderPage(route, navigator) {
-    const props = route.props || {};
-    props.navigator = navigator;
-
-    return React.createElement(route.component, route.props);
-  }
-
-  render() {
-    ons.mockStatusBar();
-    return (
-      <Navigator
-        swipeable={true}
-        renderPage={this.renderPage}
-        onPrePush={e => console.log('prepush', e)}
-        onPostPush={e => console.log('postpush', e)}
-        onPrePop={e => console.log('prepop', e)}
-        onPostPop={e => console.log('postpop', e)}
-        initialRoute={{
-          component: Examples,
-          props: {
-            key: 'examples'
-          }
-        }}
-      />
-    );
-  }
-}
-
-ReactDOM.render(<App />, document.getElementById('app'));
+export default Examples;
