+++ bindings/react/react-onsenui-examples/src/components/ActionSheet.js
