+++ bindings/react/react-onsenui-examples/src/components/AlertDialog.js
