+++ bindings/react/react-onsenui-examples/src/components/BackButton.js
