+++ bindings/react/react-onsenui-examples/src/components/BottomToolbar.js
