+++ bindings/react/react-onsenui-examples/src/components/Carousel.js
