+++ bindings/react/react-onsenui-examples/src/components/Dialog.js
