+++ bindings/react/react-onsenui-examples/src/components/GestureDetector.js
