+++ bindings/react/react-onsenui-examples/src/components/Icon.js
