+++ bindings/react/react-onsenui-examples/src/components/Input.js
