+++ bindings/react/react-onsenui-examples/src/components/LazyList.js
