+++ bindings/react/react-onsenui-examples/src/components/Modal.js
