+++ bindings/react/react-onsenui-examples/src/components/MyToolbar.js
