+++ bindings/react/react-onsenui-examples/src/components/Page.js
