+++ bindings/react/react-onsenui-examples/src/components/ProgressBar.js
