+++ bindings/react/react-onsenui-examples/src/components/Range.js
