+++ bindings/react/react-onsenui-examples/src/components/ReduxNavigator/actions.js
