+++ bindings/react/react-onsenui-examples/src/components/ReduxNavigator/index.js
