+++ bindings/react/react-onsenui-examples/src/components/ReduxNavigator/main-page.js
