+++ bindings/react/react-onsenui-examples/src/components/ReduxNavigator/my-toolbar.js
