+++ bindings/react/react-onsenui-examples/src/components/ReduxNavigator/reducer.js
