+++ bindings/react/react-onsenui-examples/src/components/ReduxNavigator/secondary-page.js
