+++ bindings/react/react-onsenui-examples/src/components/Ripple.js
