+++ bindings/react/react-onsenui-examples/src/components/RouterNavigator.js
