+++ bindings/react/react-onsenui-examples/src/components/Segment.js
