+++ bindings/react/react-onsenui-examples/src/components/SpeedDial.js
