+++ bindings/react/react-onsenui-examples/src/components/Splitter.js
