+++ bindings/react/react-onsenui-examples/src/components/Switch.js
