+++ bindings/react/react-onsenui-examples/src/components/Tabbar.js
