+++ bindings/react/react-onsenui-examples/src/components/Toast.js
