+++ bindings/react/react-onsenui-examples/src/components/index.js
